% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQ5D_Scoring_Algorithm.R
\name{eq5d}
\alias{eq5d}
\title{Calculates the U.S. preference-weighted EQ-5D-3L index score.}
\usage{
eq5d(data, varnames)
}
\arguments{
\item{data}{A data frame.}

\item{varnames}{A character vector specifying the variable names of the five
questions used to compute the EQ5D index score}
}
\value{
This function returns the original data frame provided as input with 
   an additional variable storing the EQ5D index score.
}
\description{
Calculates the U.S. preference-weighted EQ-5D-3L index score.
}
