% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOOS_Scoring_Algorithm.R
\name{hoos}
\alias{hoos}
\title{Calculates the Hip Disability and Osteoarthritis Outcomes Survey subscale scores (HOOS).}
\usage{
hoos(data, symptomVars, painVars, adlVars, sportVars, qolVars)
}
\arguments{
\item{data}{A data frame.}

\item{symptomVars}{A character vector of length 5 specifying the name of the pain variable}

\item{painVars}{A character vector of length 10 specifying the name of the limp variable}

\item{adlVars}{A character vector of length 17 specifying the name of the support variable}

\item{sportVars}{A character vector of length 4 specifying the name of the distance variable}

\item{qolVars}{A character vector of length 4 specifying the name of the stairs variable}
}
\value{
This function returns the original data frame provided as input with
   five additional variables storing each subscale score.
}
\description{
The \code{hoos} function first replaces any undefined values with NA before
computing the subscale scores. At least 50% of the items in each subscale
require a response in order for the subscale score to be calculated.
}
\examples{
df <- hoos(df, symptoms, pain, adl, sport, qol)

}
