% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHHS_Scoring_Algorithm.R
\name{mhhs}
\alias{mhhs}
\title{Calculates the modified version of the Harris Hip Score (mHHS).}
\usage{
mhhs(data, pain, limp, support, distance, stairs, shoes, sitting, transport)
}
\arguments{
\item{data}{A data frame.}

\item{pain}{A character vector of length one specifying the name of the pain variable}

\item{limp}{A character vector of length one specifying the name of the limp variable}

\item{support}{A character vector of length one specifying the name of the support variable}

\item{distance}{A character vector of length one specifying the name of the distance variable}

\item{stairs}{A character vector of length one specifying the name of the stairs variable}

\item{shoes}{A character vector of length one specifying the name of the shoes variable}

\item{sitting}{A character vector of length one specifying the name of the sitting variable}

\item{transport}{A character vector of length one specifying the name of the transport variable}
}
\value{
This function returns the original data frame provided as input with
   an additional variable storing the modified HHS score.
}
\description{
\code{mHHS} pulls item scores from an internal lookup table.
}
\examples{
df <- mhhs(df, "hhs1r", "hhs2r", "hhs3r", "hhs4r", "hhs5r", "hhs6r", "hhs7r", "hhs8r")

}
