% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PediIKDC_Scoring_Algorithm.R
\name{pedi_ikdc}
\alias{pedi_ikdc}
\title{Calculates the pediatric International Knee Documentation Committee (pedi-IKDC)
score}
\usage{
pedi_ikdc(data, vars)
}
\arguments{
\item{data}{A data frame.}

\item{vars}{A character vector of length 20 specifying the names of
the 20 questions that factor into the score.}
}
\value{
This function returns the original data frame provided as input with
an additional variable storing the pedi-IKDC score.
}
\description{
The \code{pedi_ikdc} function replaces any undefined values with NA before
computing the score. The pedi-IKDC score will not be calculated if more than
two items are missing. Per the pedi-IKDC scoring instructions, item 12 does not
factor into the overall score.
}
\examples{
df <- data.frame(id=1:50,
                 q1=sample(0:4, 50, replace = T), q2=sample(0:10, 50, replace = T),
                 q3=sample(0:10, 50, replace = T), q4=sample(0:4, 50, replace = T),
                 q5=sample(0:4, 50, replace = T), q6=sample(0:4, 50, replace = T),
                 q7=sample(0:1, 50, replace = T), q8=sample(0:1, 50, replace = T),
                 q9=sample(0:4, 50, replace = T), q10=sample(0:4, 50, replace = T),
                 q11a=sample(0:4, 50, replace = T), q11b=sample(0:4, 50, replace = T),
                 q11c=sample(0:4, 50, replace = T), q11d=sample(0:4, 50, replace = T),
                 q11e=sample(0:4, 50, replace = T), q11f=sample(0:4, 50, replace = T),
                 q11g=sample(0:4, 50, replace = T), q11h=sample(0:4, 50, replace = T),
                 q11i=sample(0:4, 50, replace = T), q12=rep(0, 50),
                 q13=sample(0:10, 50, replace = T))

vars <- names(df)[2:22]

df <- pedi_ikdc(df, vars)

}
