% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickDASH_Scoring_Algorithm.R
\name{quickdash}
\alias{quickdash}
\title{Calculates the disability/symptom section and the optional high performance
sport/music or work modules of the QuickDASH.}
\usage{
quickdash(data, varnames)
}
\arguments{
\item{data}{A data frame.}

\item{varnames}{A character vector of length eleven specifying the names of
the eleven questions used to calculate the disability/symptom score.}
}
\value{
This function returns the original data frame provided as input with
an additional variable storing the QuickDASH score.
}
\description{
Calculates the disability/symptom section and the optional high performance
sport/music or work modules of the QuickDASH.
}
\examples{
df <- data.frame(id = 1:200,
q1 = sample(1:5, 200, replace = T), q2 = sample(1:5, 200, replace = T),
q3 = sample(1:5, 200, replace = T), q4 = sample(1:5, 200, replace = T),
q5 = sample(1:5, 200, replace = T), q6 = sample(1:5, 200, replace = T),
q7 = sample(1:5, 200, replace = T), q8 = sample(1:5, 200, replace = T),
q9 = sample(1:5, 200, replace = T), q10 = sample(1:5, 200, replace = T),
q11 = sample(1:5, 200, replace = T))

df <- quickdash(df, names(df)[2:11])

}
