% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOOS_Scoring_Algorithm.R
\name{hoos}
\alias{hoos}
\title{Calculates the Hip Disability and Osteoarthritis Outcomes Survey subscale scores (HOOS).}
\usage{
hoos(data, symptomvars, painvars, adlvars, sportvars, qolvars)
}
\arguments{
\item{data}{A data frame}

\item{symptomvars}{A character vector of length 5 specifying the names of the
symptom variables (which include symptoms and stiffness)}

\item{painvars}{A character vector of length 10 specifying the names of the
pain variables}

\item{adlvars}{A character vector of length 17 specifying the names of the
function in activities of daily living variables}

\item{sportvars}{A character vector of length 4 specifying the names of the
function in sport and recreation variables}

\item{qolvars}{A character vector of length 4 specifying the names of the
hip-related quality of life variables}
}
\value{
This function returns the original data frame provided as input with
   five additional variables storing each subscale score.
}
\description{
At least half of the items in each subscale require a response in order for
that subscale score to be calculated.
}
\examples{
fai <- read.csv("FAI test data.csv", stringsAsFactors = F)

hoosvars <- names(fai)[c(31:109)]
symptoms <- names(fai)[seq(31, 39, 2)]
pain <- names(fai)[c(41, 59, 2)]
adl <- names(fai)[c(61, 93, 2)]
sport <- names(fai)[c(95, 101, 2)]
qol <- names(fai)[c(103, 109, 2)]

library(dplyr)
test <- cbind.data.frame(select(fai, -(hoossy1r:hoosq4l)),
                         apply(fai[hoosvars],
                               2,
                               function(x) ifelse(x \%in\% c(1,2,3,4,5), x-1, x)))

df <- hoos(df, symptoms, pain, adl, sport, qol)

}
