% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickDASH_Scoring_Algorithm.R
\name{quickdash}
\alias{quickdash}
\title{Calculates the QuickDASH score}
\usage{
quickdash(data, symptomvars, include_work = F, workvars = NULL,
  include_sports = F, sportsvars = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{symptomvars}{A character vector of length eleven specifying the names of
the eleven questions used to calculate the disability/symptom score.}

\item{include_work}{Logical value specifying whether to calculate a score for
the optional work module}

\item{workvars}{An optional character vector of length four specifying the
names of the four questions used to score the work module.}

\item{include_sports}{Logical value specifying whether to calculate a score for
the optional sports/performing arts module}

\item{sportsvars}{An optional character vector of length four specifying the
names of the four questions used to score the sports/performing arts module.}
}
\value{
This function returns the original data frame provided as input with
an additional variable storing the QuickDASH score. Additional variables
storing the optional module scores are created upon request.
}
\description{
Calculates the disability/symptom QuickDASH score, and the
optional high performance sport/music or work module scores when specified.
The QuickDASH score will \strong{not} be calculated if more than 1 item is
missing. An optional module score will \strong{not} be calculated if any
of the items are missing.
}
\examples{
df <- data.frame(id = 1:200, replicate(15, sample(1:5, 200, replace = T)))

## disability/symptom score only
df <- quickdash(df, names(df)[2:12])

## optional work module score
df <- quickdash(df, names(df)[2:12], include_work=T, workvars=names(df)[13:16])

}
